Fix issue with loading a saved fm_taskTable object
Background for this issue is given in the Mathworks article "Avoid
Property Initialization Order Dependency" and context for it is in "Save
and Load Process for Objects". In essence, when loading an object from a
saved .mat file, the object has to reinitialize. During this process,
saved property values are copied into the object's property fields and
then all the saved property values that have an associated set method are
called with the corresponding saved property values (thus the set method
code is run again). The set methods might called in a random order, which
means that if a set method running correctly is dependent on another set
method having run first, the user might encounter an unexpected issue.