Fix fm_taskTable failing with only 1 item
fm_taskTable.convertToNumerical() used cell2table, which would
concatenate columns if it could, so when there was 1 row in the table
(or multiple rows that all have the same dimensions), the table's
entries would not be enclosed in a cell array, which is inconsistent
with their usual and necessary behavior. Doubly enclosing contents
in a cell array fixed the issue because 1-element cell arrays are always
concatenaable.