Add validation & table expansion to fm_task
fm_task (previously named fm_taskTable) now has all the features it
needs for easy MVPA simulation. It now checks and corrects inputs
and has an "expand" option that will take a "template" task and
extrapolate all the rows necessary to create independent variation
of conditions. This saves the user from needing to calculate these
manually.

I'll elaborate on more of this functionality when I write the
helps.