Clean up functionalMage.m, rename to funkyMage.m
Initially, the plan was to make this class (func*Mage.m) be a wrapper for
all the basic functionality offered by this toolbox. However, that might
hide the flow of the simulation and analysis process too much. But if
there's need for it I'll go back to the initial plan.